+++ main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "status_led.cpp" "main.cpp" "connection_manager.cpp"
+idf_component_register(SRCS "visitor.cpp" "status_led.cpp" "main.cpp" "connection_manager.cpp"
                     INCLUDE_DIRS "include")
